pyproject.toml
[tool.poetry]
name = "repl_python3_DatteBot-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.17.0"
flask = "^1.1.2"
youtube-dl = "^2021.4.17"
ffmpeg = "^1.4"
PyNaCl = "^1.4.0"
prsaw = "^0.3.1"
aiohttp-apiset = "^0.9.13"
aiohttp-basicauth = "^0.2.0"
aiohttp-utils = "^3.1.1"
google-api-python-client = "^2.12.0"
bs4 = "^0.0.1"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
